edit card component Relates #19
